+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Boost 1.77.0:
 
+*  Implicit conversion operator from `string` to `std::string_view`.
 * `object` deallocates the correct size.
 
 Boost 1.76.0: